Implemented Singleton & Null Object
Implemented Singleton and Null Object design patterns by creating a nested class within (and used only by) the Memento class. The NullInstance is forced to exist only once, is immutable, and contains non-impacting attributes, satisfying the requirements of both the Singleton and Null Object design patterns. Also added some commenting to the remainder of the Memento class and updated the ReadMe.