Visual Changes for GUI
Made multiple changes as indicated by the client about the layout of the GUI class representation. Classes are now represented by rectangles, with fields/methods/parameters shown in traditional UML style and a rectangle that resizes as fields are added. Also removed Obsolete folder, so it doesn't mess with our code coverage.